+++ src/afl-fuzz-init.c
@@ -3496,13 +3496,8 @@ void setup_testcase_shmem(afl_state_t *afl) {
   setenv(SHM_FUZZ_MAP_SIZE_ENV_VAR, shm_fuzz_map_size_str, 1);
   ck_free(shm_fuzz_map_size_str);
 
-#ifdef USEMMAP
-  setenv(SHM_FUZZ_ENV_VAR, afl->shm_fuzz->g_shm_file_path, 1);
-#else
-  u8 *shm_str = alloc_printf("%d", afl->shm_fuzz->shm_id);
-  setenv(SHM_FUZZ_ENV_VAR, shm_str, 1);
-  ck_free(shm_str);
-#endif
+  afl_shm_fuzz_env_set(afl->shm_fuzz);
+
   afl->fsrv.support_shmem_fuzz = 1;
   afl->fsrv.shmem_fuzz_len = (u32 *)map;
   afl->fsrv.shmem_fuzz = map + sizeof(u32);